+++ MariCommands/Services/CommandService.cs
@@ -1,27 +1,23 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
-using MariGlobals.Extensions;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 
 namespace MariCommands
 {
     /// <inheritdoc />
     public class CommandService : ICommandService
     {
-        private readonly IServiceProvider _provider;
+        private readonly IContextExecutor _contextExecutor;
+
         /// <summary>
         /// Creates an instance of <see cref="CommandService" />.
         /// </summary>
-        public CommandService(IServiceProvider provider)
+        public CommandService(IContextExecutor contextExecutor)
         {
-            _provider = provider;
+            _contextExecutor = contextExecutor;
         }
 
-        private CommandDelegate CommandDelegate { get; set; }
 
         /// <inheritdoc />
         public IModule AddModule<T>(T type) where T : IModule
@@ -49,59 +45,18 @@ namespace MariCommands
 
         /// <inheritdoc />
         public Task<IResult> ExecuteAsync(string input, CommandContext context)
-        {
-            context.RawArgs = input;
-
-            return ExecuteAsync(context);
-        }
+            => _contextExecutor.ExecuteAsync(input, context);
 
         /// <inheritdoc />
         public Task<IResult> ExecuteAsync(ICommand command, string args, CommandContext context)
-        {
-            context.Command = command;
-            context.RawArgs = args;
-
-            return ExecuteAsync(context);
-        }
+            => _contextExecutor.ExecuteAsync(command, args, context);
 
         /// <inheritdoc />
         public Task<IResult> ExecuteAsync(ICommand command, IEnumerable<object> args, CommandContext context)
-        {
-            context.Command = command;
-            context.Args = args.ToList();
-
-            return ExecuteAsync(context);
-        }
-
-        private async Task<IResult> ExecuteAsync(CommandContext context)
-        {
-            CheckInitialized();
-
-            var scope = _provider.CreateScope();
-
-            context.ServiceProvider = scope.ServiceProvider;
-            context.RegisterForDispose(scope);
-
-            await CommandDelegate(context);
-
-            return context.Result;
-        }
+            => _contextExecutor.ExecuteAsync(command, args, context);
 
         /// <inheritdoc />
         public void Initialize(CommandDelegate commandDelegate)
-        {
-            if (CommandDelegate.HasContent())
-                throw new InvalidOperationException("The current application has already started.");
-
-            commandDelegate.NotNull(nameof(commandDelegate));
-
-            CommandDelegate = commandDelegate;
-        }
-
-        private void CheckInitialized()
-        {
-            if (CommandDelegate.HasNoContent())
-                throw new InvalidOperationException("The current application has not started yet.");
-        }
+            => _contextExecutor.Initialize(commandDelegate);
     }
 }
