+++ MariCommands/Services/IContextExecutor.cs
@@ -7,7 +7,7 @@ namespace MariCommands
     /// <summary>
     /// A service that can execute commands.
     /// </summary>
-    public interface ICommandExecutor
+    public interface IContextExecutor
     {
         /// <summary>
         /// Execute a command with the specified text input and returns the result.
@@ -61,5 +61,11 @@ namespace MariCommands
         ///     <param ref="commandContext" /> must not be null.
         /// </exception>
         Task<IResult> ExecuteAsync(ICommand command, IEnumerable<object> args, CommandContext commandContext);
+
+        /// <summary>
+        /// Initialize the current command service with the specified command middleware pipeline.
+        /// </summary>
+        /// <param name="commandDelegate">The command middleware for proccess command requests.</param>
+        void Initialize(CommandDelegate commandDelegate);
     }
 }
